Add option for an additional 'misc' vault hash.
This allows the use of vault for all configuration management
and not just secrets.  Any values in this hash that match
the name of the CLI flag will be used (if not otherwise provided
on the command line).

This new hash is checked __first__, so if values happen to be
present in one of the other hashes they will be used instead.